feat(faq): make FAQSection dynamic by passing block props and updating FAQList
This commit introduces changes to make the FAQSection component dynamic by passing the entire block object as props. Additionally, the FAQList component is updated to accept a list of FAQ items and render them using PortableText. This allows for more flexible and customizable FAQ sections based on the data provided.