feat(sanity): add script to delete all documents of a specific type using Sanity CLI
This commit introduces a new script `deleteAllOfTypeWithCLI.js` that allows for the deletion of all documents of a specified type in a Sanity project using the Sanity CLI. The script is included in the `tsconfig.json` to ensure it is part of the TypeScript build process. This addition facilitates easier management and cleanup of documents in the Sanity dataset.